+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `14:03` -- **Updated** `apps/desktopProbe/src/components/linksList.tsx`
 - `14:02` -- **Updated** `apps/desktopProbe/src/components/linksList.tsx`
 - `14:02` -- **Updated** `apps/desktopProbe/src/components/linksList.tsx`
 - `14:00` -- **Updated** `libraries/core/src/types.ts`